+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.4.2 (4.10.2024)
+
+## New
+- Added `Delver's Journey` season stages (Delve progression)
+
+## Misc
+- Changed `level` and `maxLevel` tags to hide if level = maxLevel
+
 # v1.4.2 (14.9.2024)
 
 ## New